Use size_t instead of int for remote file size.
git-svn-id: svn+ssh://cvs.lpds.sztaki.hu/var/lib/svn/szdg/dcapi/trunk@2407 a7169a2c-3604-0410-bc95-c702d8d87f7a

+++ dcapi/boinc/wu.C
@@ -726,7 +726,7 @@ int DC_addWUInput(DC_Workunit *wu, const char *logicalFileName, const char *URL,
 	{
 		va_start(ap, fileMode);
 		char *md5 = va_arg(ap, char *);
-		int size = va_arg(ap, int);
+		size_t size = va_arg(ap, size_t);
 		va_end(ap);
 
 		rfile = _DC_createRemoteFile(logicalFileName, URL, md5, size);
